split role-binding commands into individual files
