+++ test/fixtures/output/iam/rbac/role-binding/list-group-and-role-with-no-matches-onprem.golden
@@ -1,2 +1,2 @@
-  Principal | Role | ResourceType | Name | PatternType  
-------------+------+--------------+------+--------------
+  Principal | Role | Resource Type | Name | Pattern Type  
+------------+------+---------------+------+---------------
