+++ test/fixtures/output/iam/rbac/role-binding/list-group-and-role-with-one-resource-onprem.golden
@@ -1,3 +1,3 @@
-    Principal   |      Role      | ResourceType |  Name  | PatternType  
-----------------+----------------+--------------+--------+--------------
-  Group:hobbits | DeveloperWrite | Topic        | shire- | PREFIXED     
+    Principal   |      Role      | Resource Type |  Name  | Pattern Type  
+----------------+----------------+---------------+--------+---------------
+  Group:hobbits | DeveloperWrite | Topic         | shire- | PREFIXED
