+++ test/fixtures/output/iam/rbac/role-binding/list-user-and-role-with-cluster-resource-onprem.golden
@@ -1,3 +1,3 @@
-  Principal  |     Role      | ResourceType | Name | PatternType  
--------------+---------------+--------------+------+--------------
-  User:frodo | SecurityAdmin | Cluster      |      |              
+  Principal  |     Role      | Resource Type | Name | Pattern Type  
+-------------+---------------+---------------+------+---------------
+  User:frodo | SecurityAdmin | Cluster       |      |
