+++ test/fixtures/output/iam/rbac/role-binding/list-user-and-role-with-multiple-resources-from-one-group-onprem.golden
@@ -1,4 +1,4 @@
-    Principal   |     Role      | ResourceType | Name  | PatternType  
-----------------+---------------+--------------+-------+--------------
-  Group:hobbits | DeveloperRead | Topic        | drink | LITERAL      
-  Group:hobbits | DeveloperRead | Topic        | food  | LITERAL      
+    Principal   |     Role      | Resource Type | Name  | Pattern Type  
+----------------+---------------+---------------+-------+---------------
+  Group:hobbits | DeveloperRead | Topic         | drink | LITERAL       
+  Group:hobbits | DeveloperRead | Topic         | food  | LITERAL
