+++ test/fixtures/output/iam/rbac/role-binding/list-user-and-role-with-resources-from-multiple-groups-onprem.golden
@@ -1,4 +1,4 @@
-      Principal     |      Role      | ResourceType |  Name  | PatternType  
---------------------+----------------+--------------+--------+--------------
-  Group:hobbits     | DeveloperWrite | Topic        | shire- | PREFIXED     
-  Group:ringBearers | DeveloperWrite | Topic        | ring-  | PREFIXED     
+      Principal     |      Role      | Resource Type |  Name  | Pattern Type  
+--------------------+----------------+---------------+--------+---------------
+  Group:hobbits     | DeveloperWrite | Topic         | shire- | PREFIXED      
+  Group:ringBearers | DeveloperWrite | Topic         | ring-  | PREFIXED
