+++ test/fixtures/output/iam/rbac/role-binding/list-user-onprem.golden
@@ -1,7 +1,7 @@
-      Principal     |      Role      | ResourceType |  Name  | PatternType  
---------------------+----------------+--------------+--------+--------------
-  Group:hobbits     | DeveloperRead  | Topic        | drink  | LITERAL      
-  Group:hobbits     | DeveloperRead  | Topic        | food   | LITERAL      
-  Group:hobbits     | DeveloperWrite | Topic        | shire- | PREFIXED     
-  Group:ringBearers | DeveloperWrite | Topic        | ring-  | PREFIXED     
-  User:frodo        | SecurityAdmin  | Cluster      |        |              
+      Principal     |      Role      | Resource Type |  Name  | Pattern Type  
+--------------------+----------------+---------------+--------+---------------
+  Group:hobbits     | DeveloperRead  | Topic         | drink  | LITERAL       
+  Group:hobbits     | DeveloperRead  | Topic         | food   | LITERAL       
+  Group:hobbits     | DeveloperWrite | Topic         | shire- | PREFIXED      
+  Group:ringBearers | DeveloperWrite | Topic         | ring-  | PREFIXED      
+  User:frodo        | SecurityAdmin  | Cluster       |        |
